+++ iShangkeProject/iShangkeProject-Service/src/main/java/com/ishangke/edunav/manager/impl/UserManagerImpl.java
@@ -244,6 +244,12 @@ public class UserManagerImpl implements UserManager {
         if (isValid) {
             throw new ManagerException("Register user with phone number: " + userBo.getPhone() + " failed because authCode: " + sessionBo.getAuthCode() + "does not match");
         }
+        if (userBo.getPhone() == null) {
+            throw new ManagerException("User registration must specify phone");
+        }
+        if (userBo.getPassword() == null) {
+            throw new ManagerException("User registration must specify password");
+        }
         
         // 判断是否存在手机号码一样的USER
         UserEntityExt entityInDb = new UserEntityExt();
@@ -269,6 +275,13 @@ public class UserManagerImpl implements UserManager {
         }
         else {
             throw new AuthenticationException("Non-admin user trying createUser");
+        }
+        
+        if (targetUser.getPhone() == null) {
+            throw new ManagerException("User creation must specify phone");
+        }
+        if (targetUser.getPassword() == null) {
+            throw new ManagerException("User creation must specify password");
         }
         
         // 判断是否存在手机号码一样的USER
@@ -300,6 +313,24 @@ public class UserManagerImpl implements UserManager {
         if (partner.getId() <= 0) {
             throw new AuthenticationException("Partner Id not specified at createPartneruser");
         }
+        if (targetUser.getReference() == null) {
+            throw new ManagerException("PartnerUser creation must specify reference");
+        }
+        if (targetUser.getPassword() == null) {
+            throw new ManagerException("PartnerUser creation must specify password");
+        }
+        
+        
+        if (targetUser.getPhone() != null) {
+           // 判断是否存在手机号码一样的USER
+           UserEntityExt entityInDb = new UserEntityExt();
+           entityInDb.setPhone(targetUser.getPhone());
+
+           List<UserEntityExt> entityList = userMapper.list(entityInDb, null);
+           if (entityList != null && entityList.size() != 0) {
+               throw new ManagerException(targetUser.getPhone() + " is already in db");
+           }
+       }
         
         UserBo response = null;
         try {
@@ -341,7 +372,7 @@ public class UserManagerImpl implements UserManager {
             throw new ManagerException("OpenCellSession with phone number: " + userBo.getPhone() + " failed", t);
         }
         
-        if (result != null && result.size() >= 0) {
+        if (result != null && result.size() > 0) {
             throw new ManagerException("Account already registered");
         }
         
@@ -497,6 +528,9 @@ public class UserManagerImpl implements UserManager {
         
         UserEntityExt curUser = null;
         try {
+            if (!authManager.isAbleToLogin(loginBo.getAccountIdentifier())) {
+                throw new ManagerException("User cannot login, please wait for a minute");
+            }
             UserEntityExt search = new UserEntityExt();
             search.setPhone(loginBo.getAccountIdentifier());
             
@@ -514,9 +548,11 @@ public class UserManagerImpl implements UserManager {
             userMapper.update(curUser);
             
         } catch (Throwable t) {
+            authManager.fail(loginBo.getAccountIdentifier());
             throw new ManagerException("LoginByPhone failed for user: " + loginBo.getAccountIdentifier(), t);
         }
         
+        authManager.success(loginBo.getAccountIdentifier());
         return UserConverter.toBo(curUser);
     }
 
@@ -528,6 +564,9 @@ public class UserManagerImpl implements UserManager {
         
         UserEntityExt curUser = null;
         try {
+            if (!authManager.isAbleToLogin(loginBo.getAccountIdentifier())) {
+                throw new ManagerException("User cannot login, please wait for a minute");
+            }
             UserEntityExt search = new UserEntityExt();
             search.setReference(loginBo.getAccountIdentifier());
             
@@ -545,9 +584,11 @@ public class UserManagerImpl implements UserManager {
             userMapper.update(curUser);
             
         } catch (Throwable t) {
+            authManager.fail(loginBo.getAccountIdentifier());
             throw new ManagerException("LoginByReference failed for user: " + loginBo.getAccountIdentifier(), t);
         }
         
+        authManager.success(loginBo.getAccountIdentifier());
         return UserConverter.toBo(curUser);
     }
 
@@ -570,7 +611,9 @@ public class UserManagerImpl implements UserManager {
         else {
             throw new AuthenticationException("Non-admin user trying deleting someone else's user");
         }
-        
+        if (targetUserEntity.getId() == null) {
+            throw new ManagerException("User deletion must specify id");
+        }
         try {
             targetUserEntity.setDeleted(1);
             userMapper.deleteById(targetUserEntity.getId());
@@ -600,7 +643,12 @@ public class UserManagerImpl implements UserManager {
             }
         }
         
+        if (targetUserEntity.getId() == null) {
+            throw new ManagerException("User udpate must specify id");
+        }
         targetUserEntity.setLastModifyTime(DateUtility.getCurTimeInstance());
+        targetUserEntity.setCreateTime(null);
+        targetUserEntity.setDeleted(null);
         try {
             userMapper.update(targetUserEntity);
         } catch (Throwable t) {
